requirements.txt
tweepy==4.12.1
numpy==1.26.2
pandas==2.1.4
plotly==5.13.0
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
textblob==0.17.1
matplotlib==3.8.2
seaborn==0.13.0
textblob==0.17.1